+++ Test/Dec.hs
@@ -39,6 +39,11 @@ $(S.dectest11)
 #endif
 $(S.dectest12)
 $(S.dectest13)
+$(S.dectest14)
+
+#if __GLASGOW_HASKELL__ >= 710
+$(S.dectest15)
+#endif
 
 $(fmap unqualify S.instance_test)